fixed game interrupted bug
