+++ frontend/src/index.html
@@ -490,7 +490,7 @@
 
     <section id="Game">
         <h1 class="title">PONG!</h1>
-        <pong-game></pong-game>
+        <pong-remote></pong-remote>
         <score-board class="d-none"></score-board>
     </section>
