+++ frontend/src/js/customElements/PongAi.js
@@ -1,47 +1,27 @@
-import BaseElement from "./BaseElement.js";
+import Pong from "./Pong.js";
 
-class PongAi extends BaseElement {
+class PongAi extends Pong {
   constructor() {
     super();
-    this.init();
     this.animationFrameId = null;
     this.lastTime = performance.now();
   }
 
   init() {
-    this.innerHTML = "";
-    this.canvas = document.createElement("canvas");
-    this.canvas.width = 1000;
-    this.canvas.height = 700;
-    this.scoreboard = this.createElement("scoreboard");
-    this.side1Username = this.createElement("side1Username");
-    this.side2Username = this.createElement("side2Username");
-    this.readyButton = document.createElement("button");
-    this.readyButton.id = "readyButton";
-    this.readyButton.textContent = "Ready to Play";
-
-    this.append(
-      this.canvas,
-      this.scoreboard,
-      this.side1Username,
-      this.side2Username,
-      this.readyButton
-    );
-
+    super.init();
+    
     this.readyButton.addEventListener("click", () => {
       this.readyButton.style.display = "none";
       this.lastTime = performance.now();
       this.gameLoop(this.lastTime);
     });
 
-    this.ctx = this.canvas.getContext("2d");
-
-    // Game state
     const paddleHeight = this.canvas.height * 0.25;
     const paddleWidth = this.canvas.width * 0.02;
     const paddleInitialY = (this.canvas.height - paddleHeight) / 2;
     const playerPaddleX = 0;
     const aiPaddleX = this.canvas.width - paddleWidth;
+    
     this.playerPaddle = {
       x: playerPaddleX,
       y: paddleInitialY,
@@ -49,6 +29,7 @@ class PongAi extends BaseElement {
       height: paddleHeight,
       speed: 6,
     };
+    
     this.aiPaddle = {
       x: aiPaddleX,
       y: paddleInitialY,
@@ -56,50 +37,20 @@ class PongAi extends BaseElement {
       height: paddleHeight,
       speed: 7,
     };
+    
     this.ball = { x: 450, y: 250, size: 10, vx: 3, vy: 3 };
-
-    this.gameOver = false;
-
-    this.keys = {};
     this.aiKeys = { ArrowUp: false, ArrowDown: false };
-
     this.aiInterval = null;
-    this.animationFrameId = null;
-
-    this.handleKeyDown = this.handleKeyDown.bind(this);
-    this.handleKeyUp = this.handleKeyUp.bind(this);
+    this.firstPrediction = false;
     this.gameLoop = this.gameLoop.bind(this);
   }
 
-  createElement(id) {
-    const el = document.createElement("span");
-    el.id = id;
-    return el;
-  }
-
-  setPositions({
-    leftPaddle = { top: "calc(50% - 50px)" },
-    rightPaddle = { top: "calc(50% - 50px)" },
-    ball = { top: "calc(50% - 5px)", left: "calc(50% - 5px)" },
-  }) {
-    Object.assign(this.paddels.left.style, leftPaddle);
-    Object.assign(this.paddels.right.style, rightPaddle);
-    Object.assign(this.ball.style, ball);
-  }
-
   startGame(difficulty) {
-    this.addEventListeners();
-    this.page.app.stateManager.updateState("currentGame", true);
+    super.startGame();
     this.resetBall();
-    this.playerScore = this.aiScore = 0;
-    this.updateScoreDisplay();
-    this.gameOver = false;
-
     this.ctx.clearRect(0, 0, this.canvas.width, this.canvas.height);
-
     const diff = difficulty ? difficulty.toLowerCase() : "easy";
-    this.side1Username.textContent = this.page.app.auth.user.username;
-    this.side2Username.textContent = "AI" + (difficulty ? ` (${difficulty})` : "");
+    this.setSideUsernames(this.page.app.auth.user.username, "AI" + (difficulty ? ` (${difficulty})` : ""));
 
     if (diff === "hard") {
       this.aiPaddle.speed = 7;
@@ -227,7 +178,7 @@ class PongAi extends BaseElement {
           }
         }
       } else {
-        this.simulateAIInput(this.difficulty);
+        this.simulateAIInput(this.selectedDifficulty);
       }
     }
 
@@ -275,9 +226,9 @@ class PongAi extends BaseElement {
     }
 
     if (this.ball.x <= 0) {
-      this.aiScore++;
+      this.opponentScore++;
       this.updateScoreDisplay();
-      if (this.aiScore >= 3) {
+      if (this.opponentScore >= 3) {
         this.endGame();
       } else {
         this.resetBall();
@@ -293,15 +244,10 @@ class PongAi extends BaseElement {
     }
   }
 
-  updateScoreDisplay() {
-    this.scoreboard.textContent = `${this.playerScore} - ${this.aiScore}`;
-  }
-
   draw() {
     this.ctx.clearRect(0, 0, this.canvas.width, this.canvas.height);
     this.ctx.fillStyle = "#202428";
     this.ctx.fillRect(0, 0, this.canvas.width, this.canvas.height);
-
     this.ctx.fillStyle = "#f1f1f1";
     this.ctx.fillRect(
       this.playerPaddle.x,
@@ -328,53 +274,13 @@ class PongAi extends BaseElement {
     this.firstPrediction = false;
   }
 
-  handleKeyDown(e) {
-    this.keys[e.key.toLowerCase()] = true;
-  }
-
-  handleKeyUp(e) {
-    this.keys[e.key.toLowerCase()] = false;
-  }
-
-  endGame(message) {
-    this.cleanup();
-    const gameScore = {
-      playerScore: this.playerScore,
-      aiScore: this.aiScore,
-    };
-
-    this.dispatchEvent(
-      new CustomEvent("gameEnd", {
-        detail: gameScore,
-        bubbles: true,
-        composed: true,
-      })
-    );
-  }
-
-  addEventListeners() {
-    window.addEventListener("keydown", this.handleKeyDown);
-    window.addEventListener("keyup", this.handleKeyUp);
-  }
-
   cleanup() {
-    window.removeEventListener("keydown", this.handleKeyDown);
-    window.removeEventListener("keyup", this.handleKeyUp);
+    super.cleanup();
     if (this.animationFrameId) cancelAnimationFrame(this.animationFrameId);
     if (this.aiInterval) clearInterval(this.aiInterval);
-    this.gameOver = true;
-    if (this.page?.app.stateManager.state.currentGame) {
-      this.page?.app.stateManager.updateState("currentGame", false);
-    }
-  }
-
-  connectedCallback() {
-    this.addEventListeners();
-  }
-
-  disconnectedCallback() {
-    this.cleanup();
   }
 }
 
 customElements.define("pong-ai", PongAi);
+
+export default PongAi;
